requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
langchain==0.0.350
sentence-transformers==2.2.2
pymupdf==1.23.8
faiss-cpu==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0
